+++ Assets/_Game/Features/Gameplay/Scripts/DrawPhase/IDrawPhaseModel.cs
@@ -1,6 +1,7 @@
-using System;
 using UnityObservables;
 
-public interface IDrawPhaseModel : IDisposable{
-    void Initialize();
+public interface IDrawPhaseModel : IModel
+{
+    Observable<int> CardsToDraw { get; }
+    Observable<bool> IsDrawing { get; }
 }