+++ Assets/_Game/Features/Gameplay/Scripts/FusePhase/IFusePhaseController.cs
@@ -1,5 +1,6 @@
-using System;
-
-public interface IFusePhaseController : IDisposable{
-    void Initialize();
+public interface IFusePhaseController : IController
+{
+    void SetUnits(string primaryId, string secondaryId);
+    void Fuse();
+    void Cancel();
 }