+++ Assets/_Game/Features/Gameplay/Scripts/FusePhase/IFusePhaseModel.cs
@@ -1,6 +1,8 @@
-using System;
 using UnityObservables;
 
-public interface IFusePhaseModel : IDisposable{
-    void Initialize();
+public interface IFusePhaseModel : IModel
+{
+    Observable<bool> IsActive { get; }
+    Observable<string> PrimaryUnitId { get; }
+    Observable<string> SecondaryUnitId { get; }
 }