+++ Assets/_Game/Features/Gameplay/Scripts/GameState/IGameStateSubsystem.cs
@@ -1,11 +1,5 @@
-using UnityEngine.Events;
-
 public interface IGameStateSubsystem : ISubsystem
 {
-    event UnityAction<int> CurrentTurnChanged;
-    event UnityAction<string> CurrentPhaseChanged;
-    event UnityAction<int> MatchTimerChanged;
-
-    void StartMatch();
-    void EndTurn();
+    IGameStateModel Model { get; }
+    IGameStateController Controller { get; }
 }