+++ Assets/_Game/Features/Gameplay/Scripts/Hand/IHandController.cs
@@ -1,5 +1,6 @@
-using System;
+using System.Threading.Tasks;
 
-public interface IHandController : IDisposable{
-    void Initialize();
+public interface IHandController : IController
+{
+    Task PlayCard(string cardId);
 }